soluciona problema cuando no se encuentran links en archivos y directorios
